Move `useOutsideClick` to tsx
