+++ dashboard/src/components/Header/Menu.tsx
@@ -7,6 +7,7 @@ import { CdsToggle } from "@cds/react/toggle";
 import actions from "actions";
 import { getThemeFile } from "components/HeadManager/HeadManager";
 import Row from "components/Row";
+import useOutsideClick from "components/hooks/useOutsideClick";
 import operatorIcon from "icons/olm-icon-white.svg";
 import { useEffect, useRef, useState } from "react";
 import { Helmet } from "react-helmet";
@@ -19,7 +20,6 @@ import { ThunkDispatch } from "redux-thunk";
 import { SupportedThemes } from "shared/Config";
 import { IStoreState } from "shared/types";
 import { app } from "shared/url";
-import useOutsideClick from "../js/hooks/useOutsideClick/useOutsideClick";
 import "./Menu.css";
 
 export interface IContextSelectorProps {
